+++ packages/client/.gitignore
@@ -10,6 +10,8 @@ lerna-debug.log*
 node_modules
 dist
 dist-ssr
+ssr-dist
+tsconfig.tsbuildinfo
 *.local
 
 # Editor directories and files
